requirements.txt
Django==1.9.1
django-bootstrap-form==3.2
django-filter==0.12.0
django-formtools==1.0
django-registration==2.0.4
djangorestframework==3.3.2
jsonfield==1.0.3
Markdown==2.6.5
pathspec==0.3.3
psycopg2==2.6.1
python-dateutil==2.4.2
requests==2.6.2
psycopg2==2.6.1
wheel==0.24.0
